Make sure the in mem tests run in browser and node
